update: print new tracks information after processing

+++ composeApp/src/jvmMain/kotlin/com/amarildo/spotifyfilter/ui/view/SpotifyFilter.kt
@@ -79,7 +79,7 @@ fun SpotifyFilter(vm: MyViewModel = viewModel()) {
             uiState.successMessage?.let { msg ->
                 snackbarHostState.showSnackbar(
                     message = msg,
-                    duration = SnackbarDuration.Short,
+                    duration = SnackbarDuration.Long,
                 )
                 vm.clearMessage()
             }
